feat(android): add v8 heap tool to performance
