+++ android/sdk/src/main/jni/src/bridge/entry.cc
@@ -69,24 +69,6 @@ REGISTER_JNI("com/tencent/mtt/hippy/bridge/HippyBridgeImpl", // NOLINT(cert-err5
              "(JZLcom/tencent/mtt/hippy/bridge/NativeCallback;)V",
              DestroyInstance)
 
-// [Heap]
-REGISTER_JNI("com/tencent/mtt/hippy/bridge/HippyBridgeImpl", // NOLINT(cert-err58-cpp)
-             "getHeapCodeStatistics",
-             "(J)Lcom/tencent/mtt/hippy/bridge/heap/HeapCodeStatistics;",
-             GetHeapCodeStatistics)
-REGISTER_JNI("com/tencent/mtt/hippy/bridge/HippyBridgeImpl", // NOLINT(cert-err58-cpp)
-             "getHeapStatistics",
-             "(J)Lcom/tencent/mtt/hippy/bridge/heap/HeapStatistics;",
-             GetHeapStatistics)
-REGISTER_JNI("com/tencent/mtt/hippy/bridge/HippyBridgeImpl", // NOLINT(cert-err58-cpp)
-             "getHeapSpaceStatisticsList",
-             "(J)[Lcom/tencent/mtt/hippy/bridge/heap/HeapSpaceStatistics;",
-             GetHeapSpaceStatisticsList)
-REGISTER_JNI("com/tencent/mtt/hippy/bridge/HippyBridgeImpl", // NOLINT(cert-err58-cpp)
-             "writeHeapSnapshot",
-             "(JLjava/lang/String;)I",
-             WriteHeapSnapshot)
-
 using unicode_string_view = tdf::base::unicode_string_view;
 using u8string = unicode_string_view::u8string;
 using RegisterMap = hippy::base::RegisterMap;
@@ -616,148 +598,6 @@ void DestroyInstance(__unused JNIEnv* j_env,
     }
   }
   TDF_BASE_DLOG(INFO) << "destroy end";
-}
-
-// [Heap] GetHeapCodeStatistics
-jobject GetHeapCodeStatistics(__unused JNIEnv* j_env,
-                             __unused jobject j_object,
-                             jlong j_runtime_id) {
-  TDF_BASE_DLOG(INFO) << "GetHeapCodeStatistics begin, j_runtime_id = " << j_runtime_id;
-  auto runtime_id = static_cast<int32_t>(j_runtime_id);
-  std::shared_ptr<Runtime> runtime = Runtime::Find(runtime_id);
-  v8::HeapCodeStatistics* heap_code_statistics = new v8::HeapCodeStatistics();
-  if (runtime) {
-      v8::Isolate* isolate = std::static_pointer_cast<hippy::napi::V8VM>(runtime->GetEngine()->GetVM())->isolate_;
-      isolate->GetHeapCodeAndMetadataStatistics(heap_code_statistics);
-  } else {
-      TDF_BASE_DLOG(WARNING) << "GetHeapCodeStatistics, j_runtime_id invalid";
-  }
-  // set HeapCodeStatistics data
-  jclass java_class_hcs = j_env->FindClass("com/tencent/mtt/hippy/bridge/heap/HeapCodeStatistics");
-  jmethodID hcs_constructor = j_env->GetMethodID(java_class_hcs,"<init>","(JJJ)V");
-  jobject hcs_obj = j_env->NewObject(java_class_hcs,
-                                     hcs_constructor,
-                                     (long)heap_code_statistics->code_and_metadata_size(),
-                                     (long)heap_code_statistics->bytecode_and_metadata_size(),
-                                     (long)heap_code_statistics->external_script_source_size());
-  delete heap_code_statistics;
-  heap_code_statistics = nullptr;
-  TDF_BASE_DLOG(INFO) << "GetHeapCodeStatistics end";
-  return hcs_obj;
-}
-// [Heap] GetHeapSpaceStatisticsList
-jobject GetHeapSpaceStatisticsList(__unused JNIEnv* j_env,
-                                   __unused jobject j_object,
-                                   jlong j_runtime_id) {
-    TDF_BASE_DLOG(INFO) << "GetHeapSpaceStatisticsList begin, j_runtime_id = " << j_runtime_id;
-    // init ArrayList
-    jclass java_class_list = j_env->FindClass("java/util/ArrayList");
-    jmethodID list_constructor = j_env->GetMethodID(java_class_list , "<init>","()V");
-    jobject list_obj = j_env->NewObject(java_class_list, list_constructor);
-    jmethodID list_add  = j_env->GetMethodID(java_class_list,"add","(Ljava/lang/Object;)Z");
-
-    auto runtime_id = static_cast<int32_t>(j_runtime_id);
-    std::shared_ptr<Runtime> runtime = Runtime::Find(runtime_id);
-
-    if (runtime) {
-        jclass java_class_hss = j_env->FindClass("com/tencent/mtt/hippy/bridge/heap/HeapSpaceStatistics");
-        jmethodID hss_constructor = j_env->GetMethodID(java_class_hss,"<init>","(Ljava/lang/String;JJJJ)V");
-        v8::Isolate* isolate = std::static_pointer_cast<hippy::napi::V8VM>(runtime->GetEngine()->GetVM())->isolate_;
-        size_t space_count = isolate->NumberOfHeapSpaces();
-        TDF_BASE_DLOG(INFO) << "GetHeapSpaceStatisticsList , space_count = " << space_count;
-        for(size_t i = 0 ; i < space_count ; i++)
-        {
-            // set HeapSpaceStatistics data
-            v8::HeapSpaceStatistics* heap_space_statistics = new v8::HeapSpaceStatistics;
-            isolate->GetHeapSpaceStatistics(heap_space_statistics, i);
-            jobject hss_obj = j_env->NewObject(java_class_hss,
-                                               hss_constructor,
-                                               j_env->NewStringUTF(heap_space_statistics->space_name()),
-                                               (long)heap_space_statistics->space_size(),
-                                               (long)heap_space_statistics->space_used_size(),
-                                               (long)heap_space_statistics->space_available_size(),
-                                               (long)heap_space_statistics->physical_space_size());
-            j_env->CallBooleanMethod(list_obj, list_add, hss_obj);
-            delete heap_space_statistics;
-            heap_space_statistics = nullptr;
-        }
-    } else {
-        TDF_BASE_DLOG(WARNING) << "GetHeapSpaceStatisticsList, j_runtime_id invalid";
-    }
-    TDF_BASE_DLOG(INFO) << "GetHeapSpaceStatisticsList end";
-    return list_obj;
-}
-// [Heap] GetHeapStatistics
-jobject GetHeapStatistics(__unused JNIEnv* j_env,
-                          __unused jobject j_object,
-                          jlong j_runtime_id) {
-    TDF_BASE_DLOG(INFO) << "GetHeapStatistics begin, j_runtime_id = " << j_runtime_id;
-    v8::HeapStatistics* heap_statistics = new v8::HeapStatistics();
-    auto runtime_id = static_cast<int32_t>(j_runtime_id);
-    std::shared_ptr<Runtime> runtime = Runtime::Find(runtime_id);
-    if (runtime) {
-        v8::Isolate* isolate = std::static_pointer_cast<hippy::napi::V8VM>(runtime->GetEngine()->GetVM())->isolate_;
-        isolate->GetHeapStatistics(heap_statistics);
-    } else {
-        TDF_BASE_DLOG(WARNING) << "GetHeapStatistics, j_runtime_id invalid";
-    }
-    // set HeapStatistics data
-    jclass java_class_hs = j_env->FindClass("com/tencent/mtt/hippy/bridge/heap/HeapStatistics");
-    jmethodID hs_constructor = j_env->GetMethodID(java_class_hs,"<init>","(JJJJJJJJJJJJJ)V");
-    jobject hs_obj = j_env->NewObject(java_class_hs,
-                                      hs_constructor,
-                                      (long)heap_statistics->total_heap_size(),
-                                      (long)heap_statistics->total_heap_size_executable(),
-                                      (long)heap_statistics->total_physical_size(),
-                                      (long)heap_statistics->total_available_size(),
-                                      (long)heap_statistics->total_global_handles_size(),
-                                      (long)heap_statistics->used_global_handles_size(),
-                                      (long)heap_statistics->used_heap_size(),
-                                      (long)heap_statistics->heap_size_limit(),
-                                      (long)heap_statistics->malloced_memory(),
-                                      (long)heap_statistics->external_memory(),
-                                      (long)heap_statistics->peak_malloced_memory(),
-                                      (long)heap_statistics->number_of_native_contexts(),
-                                      (long)heap_statistics->number_of_detached_contexts());
-    delete heap_statistics;
-    heap_statistics = nullptr;
-    TDF_BASE_DLOG(INFO) << "GetHeapStatistics end";
-    return hs_obj;
-}
-// [Heap] WriteHeapSnapshot
-jint WriteHeapSnapshot(__unused JNIEnv* j_env,
-                       __unused jobject j_object,
-                       jlong j_runtime_id,
-                       jstring j_heap_snapshot_path) {
-    TDF_BASE_DLOG(INFO) << "WriteHeapSnapshot begin, j_runtime_id = " << j_runtime_id;
-    auto runtime_id = static_cast<int32_t>(j_runtime_id);
-    std::shared_ptr<Runtime> runtime = Runtime::Find(runtime_id);
-    // error: runtime_id invalid
-    if (!runtime) {
-        TDF_BASE_DLOG(WARNING) << "WriteHeapSnapshot, j_runtime_id invalid";
-        return WRITE_HEAP_ERR_RUN;
-    }
-    const unicode_string_view heap_snapshot_path = JniUtils::ToStrView(j_env, j_heap_snapshot_path);
-    HeapSnapshotOutputStreamAdapter heap_snapshot_stream;
-    int set_file_ret = heap_snapshot_stream.SetFilePath(heap_snapshot_path);
-    // error: file_path invalid
-    if (set_file_ret) {
-        TDF_BASE_DLOG(WARNING) << "WriteHeapSnapshot, set_file_ret = " << set_file_ret;
-        return WRITE_HEAP_ERR_FILE;
-    }
-    v8::Isolate* isolate = std::static_pointer_cast<hippy::napi::V8VM>(runtime->GetEngine()->GetVM())->isolate_;
-    v8::HeapProfiler* heap_profiler = isolate->GetHeapProfiler();
-    const v8::HeapSnapshot* heap_snapshot = heap_profiler->TakeHeapSnapshot();
-    heap_snapshot->Serialize(&heap_snapshot_stream);
-    heap_profiler->DeleteAllHeapSnapshots();
-    delete heap_snapshot;
-    // error: save heapSnapshots error
-    if (heap_snapshot_stream.save_res_ != HeapSnapshotOutputStreamAdapter::sOk) {
-        TDF_BASE_DLOG(WARNING) << "WriteHeapSnapshot, save heapSnapshots = " << heap_snapshot_stream.save_res_;
-        return WRITE_HEAP_ERR_SAVE;
-    }
-    TDF_BASE_DLOG(INFO) << "WriteHeapSnapshot end";
-    return WRITE_HEAP_OK;
 }
 
 }  // namespace bridge
